requirements.txt
openai-whisper
pyaudio
tqdm